requirements.txt
matplotlib
numpy
pandas
# psycopg2
pyunpack
reformer-pytorch
scikit-learn
sympy
time-interpret # ==0.3.0
captum
tqdm
einops
peft
transformers
# install pytorch and cuda following readme